feat: product and metafield

+++ backend/src/main/java/vn/hust/omni/sale/service/product/application/model/InventoryRequest.java
@@ -3,11 +3,13 @@ package vn.hust.omni.sale.service.product.application.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class InventoryRequest {
     private long storeId;
     private int productId;
     private long locationId;
-    private long quantity;
+    private BigDecimal quantity;
 }
